Separate DSTU2 packages
