+++ data-query-tests/src/test/java/gov/va/api/health/dataquery/tests/dstu2/JoshBot.java
@@ -1,4 +1,4 @@
-package gov.va.api.health.dataquery.tests;
+package gov.va.api.health.dataquery.tests.dstu2;
 
 import static org.assertj.core.api.Assertions.assertThat;
